+++ .idea/workspace.xml
@@ -4,10 +4,7 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="0fe988ff-eb07-4efe-b73f-9cc42dd89e25" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/frontend/components/json/JsonBrowser.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/frontend/components/json/JsonBrowser.tsx" afterDir="false" />
-    </list>
+    <list default="true" id="0fe988ff-eb07-4efe-b73f-9cc42dd89e25" name="Changes" comment="" />
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -128,7 +125,7 @@
       <workItem from="1682593136050" duration="700000" />
       <workItem from="1682593859376" duration="20381000" />
       <workItem from="1682657098556" duration="904000" />
-      <workItem from="1682658046339" duration="17470000" />
+      <workItem from="1682658046339" duration="17835000" />
     </task>
     <task id="LOCAL-00001" summary="added json browser placeholder">
       <created>1682601641471</created>
@@ -165,7 +162,14 @@
       <option name="project" value="LOCAL" />
       <updated>1682854758387</updated>
     </task>
-    <option name="localTasksCounter" value="6" />
+    <task id="LOCAL-00006" summary="clear selected on items change">
+      <created>1682855542429</created>
+      <option name="number" value="00006" />
+      <option name="presentableId" value="LOCAL-00006" />
+      <option name="project" value="LOCAL" />
+      <updated>1682855542429</updated>
+    </task>
+    <option name="localTasksCounter" value="7" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -177,6 +181,7 @@
     <MESSAGE value="added dao layer" />
     <MESSAGE value="created json editor" />
     <MESSAGE value="refactoring clearTimeout" />
-    <option name="LAST_COMMIT_MESSAGE" value="refactoring clearTimeout" />
+    <MESSAGE value="clear selected on items change" />
+    <option name="LAST_COMMIT_MESSAGE" value="clear selected on items change" />
   </component>
 </project>
